+++ scripts/symbols.ld.sh
@@ -1,0 +1,33 @@
+#!/bin/sh
+
+#
+# file: symbols.ld.sh
+#
+# author: Copyright (C) 2017 Kamil Szczygiel http://www.distortec.com http://www.freddiechopin.info
+#
+# This Source Code Form is subject to the terms of the Mozilla Public License, v. 2.0. If a copy of the MPL was not
+# distributed with this file, You can obtain one at http://mozilla.org/MPL/2.0/.
+#
+
+set -e
+set -u
+
+cat << EOF
+/**
+ * \file
+ * \brief Symbols for linker script
+ *
+ * \author Copyright (C) 2017 Kamil Szczygiel http://www.distortec.com http://www.freddiechopin.info
+ *
+ * \par License
+ * This Source Code Form is subject to the terms of the Mozilla Public License, v. 2.0. If a copy of the MPL was not
+ * distributed with this file, You can obtain one at http://mozilla.org/MPL/2.0/.
+ *
+ * \warning
+ * Automatically generated file - do not edit!
+ *
+ * \date $(date +'%Y-%m-%d %H:%M:%S')
+ */
+
+$(while [ ${#} -gt 0 ]; do echo "PROVIDE(${1});"; shift; done)
+EOF